umu_run: use information from `toolmanifest.vdf` to construct umu's command
Using the information in the compatibility tool's manifest, umu executes
the tool's entry point as described by the tool with the expected runtime
environment. Umu now detects when the runtime should be used and potentially
use the correct runtime to do so.

This also enables umu to run tools other than proton, such as luxtorpeda
simply by setting `PROTONPATH=luxtorpeda` or `PROTONPATH=/path/to/not/proton`