Specify make targets explicitly in nix package
The umu-vendored make target depends on git submodules and is not needed
now that we provide pyzstd as a nix package.

Until now, it was disabled with a patch, however that broke in #330.

Instead of patching the makefile's `all` target, we can explicitly
specify the targets we want to build/install using make flags.

This is a temporary workaround until the Makefile can automatically
disable the target when the dependency is already available. Maybe a
make variable?