Refactor nix flake, based on overriding nixpkgs (#345)
* Use URL-like flake ref syntax for nixpkgs input

This is a purely cosmetic change.

* packaging/nix/flake.lock: Update

Flake lock file updates:

• Updated input 'nixpkgs':
    'github:NixOS/nixpkgs/3df3c47c19dc90fec35359e89ffb52b34d2b0e94?narHash=sha256-30sOEZ8CFK2nTTMdkhaNrfVlIi3rWTNV0Z5z%2BNmpFNI%3D' (2025-01-06)
  → 'github:NixOS/nixpkgs/9a5db3142ce450045840cc8d832b13b8a2018e0c?narHash=sha256-pUvLijVGARw4u793APze3j6mU1Zwdtz7hGkGGkD87qw%3D' (2025-01-29)

* Refactor nix flake, based on overriding nixpkgs

Now that umu-launcher is being maintained in nixpkgs, it can be used as
a starting point for the nix package here. We just need to override a
few things, such as the `src`, and handle any changes that have not yet
made it into a release build.

The opportunity has been taken to clean up that package-args API:
- Renamed `truststore` to `withTruststore` with a warning.
- `cbor2` wasn't working, warn when it is used.
- Added `withDeltaUpdates`

Breaking changes:
- The umu-launcher package in the overlay is no longer the unwrapped
  package (previously in `umu-launcher.nix`).
- The derivation attrs for various packages are re-written from scratch,
  likely breaking users' custom overrides.

Non-breaking changes:
- The umu & umu-run packages have effectively been renamed umu-launcher.
- The flake outputs for all systems supported by the nixpkgs package.
- The package version now defaults to the flake's git revision.

* Warn when nix package version is set unnecessarily

The README used to recommend overriding `version = shortRev`, however we
now use `shortRev` by default.

Notify users of this change by warning when `version` is overridden to
the version we would've used anyway.

* Add pyzstd to nix package inputs

Removed `zstd` from inputs as it is a transitive dependency.

* Add urllib3 to nix package inputs

* Specify make targets explicitly in nix package

The umu-vendored make target depends on git submodules and is not needed
now that we provide pyzstd as a nix package.

Until now, it was disabled with a patch, however that broke in #330.

Instead of patching the makefile's `all` target, we can explicitly
specify the targets we want to build/install using make flags.

This is a temporary workaround until the Makefile can automatically
disable the target when the dependency is already available. Maybe a
make variable?