umu_run: move toml config loading earlier and merge it with the environment
Since the toml config checks ensure that the values are mostly correct, we
can load them early and merge them with the rest of the environment. This
way they are treated as if they came from the CLI. This should allow umu
to download any proton version and its runtime set in the config.