umu_run: create `pfx/creation_sync_guard` file to satisfy proton requirements for partial prefixes
umu-launcher creates a partial prefix to control some aspects of its folder structure. Because
when creating a new prefix the `version` file doesn't exist yet, proton assumes there is no need
to upgrade the prefix and `upgrade_pfx()` doesn't run. This has the side-effect of the
`pfx/creation_sync_guard` file not being created and proton assumes the prefix was not created
properly. As a result it tries to back it up and continue, which also negates any kind of control
umu can have on the prefix.