umu_run: use `steam-runtime-launch-client` to re-enter an active container
this uses `steam-runtime-launch-client` to enable the dbus interface exposed
container runtimes as described [here](1) and re-enter it to run executables.

Exposing the dbus interface itself depends on `compatmanager_layer_name`
attribute of the tool's `toolmanifest.vdf` and is activated when
`STEAM_COMPAT_LAUNCHER_SERVICE` is set to the attribute's value. The bus's
"friendly" name is controlled by `STEAM_COMPAT_APP_ID`.

To make the bus name predictable, this change assumes only one container
instance should run on a prefix at any time. Thus `STEAM_COMPAT_APP_ID` is
set to the value of `STEAM_COMPAT_DATA_PATH` with slashes `/` replaced with
underscores `_`. When the command is built, umu checks whether the bus
name already exists, and enters the existing container using
`steam-runtime-launch-client` instead of creating an new one, in which it
executes proton with `runinprefix`.

[1](https://gitlab.steamos.cloud/steamrt/steam-runtime-tools/-/blob/main/docs/slr-for-game-developers.md?ref_type=heads#inserting-debugging-commands-into-the-container)