fix: ensure link and shim creation on runtime update
Validating the Steam Linux Runtime's platform directory through the pv-verify binary can take some time, and that phase will block before the umu symbolic link and the umu-shim file are created. If the validation were ever interrupted for any reason, a FileNotFoundError can be raised in a future session because the link will not exist. To mitigate that from occurring, ensure the link's creation on any event that occurs during validation.