refactor: use XDG_DATA_HOME and delay temp cleanup at runtime install
Refactors the install of the Steam Linux Runtime to first move the downloaded archive from the system's tmpfs or XDG_CACHE_HOME/umu to XDG_DATA_HOME/umu, and to delay the cleanup of the temporary directory after the post-install steps. Simplifies the code as it avoids handling the edge case where XDG_CACHE_HOME and XDG_DATA_HOME are on different file systems. Additionally, improves robustness of runtime installs because the cleanup of the old runtime, which can be 700MiB+ in size, would not block the post-install.