feat: add support for launching foreign arch executables inside a container runtime
Valve plans on using FEX as an additional mechanism to support running x86 executables on an aarch64 systems, which is evident through FEX being a new dependency for Proton and new configuration options in latest versions of pressure-vessel. To achieve that use case, Steam will be expected to configure pressure-vessel by setting STEAM_COMPAT_EMULATOR and/or STEAM_COMPAT_GRAPHICS_PROVIDER. The former is a JSON file which will configure pressure-vessel's container runtime machinery to use the emulator during setup, while the latter, also a JSON file, 'should be set if the ld.so, glibc, and graphics drivers for the emulated architectures are not available in the real root directory'.

See https://gitlab.steamos.cloud/steamrt/steam-runtime-tools/-/blob/v0.20251120.0/docs/steam-compat-tool-interface.md.