build(nix): add workaround for versionCheckHook
Nixpkgs recently added versionCheckHook to the package we override.

versionCheckHook expects `--version` to print the entire package
version, and it doesn't provide any way to override the version string
it expects without manually setting `version` in `preVersionCheck`.

Since the package version uses a different format to the build version,
the check fails. Patch it to ignore the -unstable-YY-MM-DD suffix.